+++ com.unity.render-pipelines.high-definition/Runtime/Debug/DebugDisplay.cs
@@ -216,6 +216,11 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             return lightingDebugSettings.debugLightingMode;
         }
 
+        public ShadowMapDebugMode GetDebugShadowMapMode()
+        {
+            return lightingDebugSettings.shadowDebugMode;
+        }
+
         public int GetDebugLightingSubMode()
         {
             switch (lightingDebugSettings.debugLightingMode)
@@ -308,6 +313,22 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             if (value != 0)
                 DisableNonMaterialDebugSettings();
             materialDebugSettings.SetDebugViewGBuffer(value);
+        }
+
+        public void SetFullScreenDebugMode(FullScreenDebugMode value)
+        {
+            if (lightingDebugSettings.shadowDebugMode == ShadowMapDebugMode.SingleShadow)
+                value = 0;
+            
+            fullScreenDebugMode = value;
+        }
+
+        public void SetShadowDebugMode(ShadowMapDebugMode value)
+        {
+            // When SingleShadow is enabled, we don't render full screen debug modes
+            if (value == ShadowMapDebugMode.SingleShadow)
+                fullScreenDebugMode = 0;
+            lightingDebugSettings.shadowDebugMode = value;
         }
 
         public void SetDebugLightingMode(DebugLightingMode value)
@@ -641,12 +662,12 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             {
                 displayName = "Shadow Debug Mode",
                 getter = () => (int)lightingDebugSettings.shadowDebugMode,
-                setter = value => lightingDebugSettings.shadowDebugMode = (ShadowMapDebugMode)value,
+                setter = value => SetShadowDebugMode((ShadowMapDebugMode)value),
                 autoEnum = typeof(ShadowMapDebugMode),
                 onValueChanged = RefreshLightingDebug
             });
 
-            if (lightingDebugSettings.shadowDebugMode == ShadowMapDebugMode.VisualizeShadowMap)
+            if (lightingDebugSettings.shadowDebugMode == ShadowMapDebugMode.VisualizeShadowMap || lightingDebugSettings.shadowDebugMode == ShadowMapDebugMode.SingleShadow)
             {
                 var container = new DebugUI.Container();
                 container.children.Add(new DebugUI.BoolField { displayName = "Use Selection", getter = () => lightingDebugSettings.shadowDebugUseSelection, setter = value => lightingDebugSettings.shadowDebugUseSelection = value, flags = DebugUI.Flags.EditorOnly, onValueChanged = RefreshLightingDebug });
@@ -670,7 +691,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             list.Add(new DebugUI.FloatField { displayName = "Shadow Range Max Value", getter = () => lightingDebugSettings.shadowMaxValue, setter = value => lightingDebugSettings.shadowMaxValue = value });
 
             list.Add(new DebugUI.EnumField { displayName = "Lighting Debug Mode", getter = () => (int)lightingDebugSettings.debugLightingMode, setter = value => SetDebugLightingMode((DebugLightingMode)value), autoEnum = typeof(DebugLightingMode), onValueChanged = RefreshLightingDebug });
-            list.Add(new DebugUI.EnumField { displayName = "Fullscreen Debug Mode", getter = () => (int)fullScreenDebugMode, setter = value => fullScreenDebugMode = (FullScreenDebugMode)value, enumNames = lightingFullScreenDebugStrings, enumValues = lightingFullScreenDebugValues, onValueChanged = RefreshLightingDebug });
+            list.Add(new DebugUI.EnumField { displayName = "Fullscreen Debug Mode", getter = () => (int)fullScreenDebugMode, setter = value => SetFullScreenDebugMode((FullScreenDebugMode)value), enumNames = lightingFullScreenDebugStrings, enumValues = lightingFullScreenDebugValues, onValueChanged = RefreshLightingDebug });
             switch (fullScreenDebugMode)
             {
                 case FullScreenDebugMode.PreRefractionColorPyramid: