+++ com.unity.render-pipelines.high-definition/Runtime/Debug/LightingDebug.cs
@@ -15,7 +15,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         IndirectDiffuseOcclusion,
         IndirectSpecularOcclusion,
         ScreenSpaceTracingRefraction,
-        ScreenSpaceTracingReflection
+        ScreenSpaceTracingReflection,
     }
 
     [GenerateHLSL]
@@ -42,11 +42,13 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         LinearSampledColor
     }
 
+    [GenerateHLSL]
     public enum ShadowMapDebugMode
     {
         None,
         VisualizeAtlas,
-        VisualizeShadowMap
+        VisualizeShadowMap,
+        SingleShadow,
     }
 
     [Serializable]
@@ -54,7 +56,12 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
     {
         public bool IsDebugDisplayEnabled()
         {
-            return debugLightingMode != DebugLightingMode.None || overrideSmoothness || overrideAlbedo || overrideNormal || overrideSpecularColor;
+            return debugLightingMode != DebugLightingMode.None
+                || overrideSmoothness
+                || overrideAlbedo
+                || overrideNormal
+                || overrideSpecularColor
+                || shadowDebugMode == ShadowMapDebugMode.SingleShadow;
         }
 
         public bool IsDebugDisplayRemovePostprocess()